Delivers data and response on successful HTTP URL Request
Note Me: And adding for Fixing issues expectedFulfillmentCount from comment,
And the XCTestExpectation.expectedFulfillmentCount` is `1` by default - so it can only be fulfilled once (unless you increase the expectedFulfillmentCount - but in our case, we want the count to be `1` because the request should complete only once!).
So since we call `fulfill` inside the URLProtocolStub request observer, and the observer is called for *every* request, we need to remove the observer after each test to make sure it's not called for a request of another test!